feature: 자기 opinion이 아니거나 admin이 아니면 수정불가능하게 만듬.

+++ module-api/src/main/java/partner42/moduleapi/service/opinion/OpinionService.java
@@ -13,8 +13,13 @@ import partner42.moduleapi.dto.opinion.OpinionUpdateRequest;
 import partner42.moduleapi.mapper.OpinionMapper;
 import partner42.modulecommon.domain.model.article.Article;
 import partner42.modulecommon.domain.model.opinion.Opinion;
+import partner42.modulecommon.domain.model.user.Role;
+import partner42.modulecommon.domain.model.user.RoleEnum;
+import partner42.modulecommon.domain.model.user.User;
+import partner42.modulecommon.domain.model.user.UserRole;
 import partner42.modulecommon.exception.ErrorCode;
 import partner42.modulecommon.exception.NoEntityException;
+import partner42.modulecommon.exception.NotAuthorException;
 import partner42.modulecommon.repository.article.ArticleRepository;
 import partner42.modulecommon.repository.member.MemberRepository;
 import partner42.modulecommon.repository.opinion.OpinionRepository;
@@ -49,12 +54,16 @@ public class OpinionService {
     }
 
     @Transactional
-    public OpinionOnlyIdResponse updateOpinion(OpinionUpdateRequest request, String opinionId) {
+    public OpinionOnlyIdResponse updateOpinion(OpinionUpdateRequest request, String opinionId, String username) {
+        verifyAuthorOfOpinion(username, opinionId);
+
         Opinion opinion = opinionRepository.findByApiId(opinionId)
             .orElseThrow(() -> new NoEntityException(ErrorCode.ENTITY_NOT_FOUND));
         opinion.updateContent(request.getContent());
         return opinionMapper.entityToOpinionOnlyIdResponse(opinion);
     }
+
+
 
     public ListResponse<OpinionResponse> findAllOpinionsInArticle(String articleId) {
         List<OpinionResponse> opinionResponses = opinionRepository.findAllByArticleApiIdAndIsDeletedIsFalse(articleId).stream()
@@ -69,7 +78,8 @@ public class OpinionService {
     }
 
     @Transactional
-    public OpinionOnlyIdResponse recoverableDeleteOpinion(String opinionId) {
+    public OpinionOnlyIdResponse recoverableDeleteOpinion(String opinionId, String username) {
+        verifyAuthorOfOpinion(username, opinionId);
         Opinion opinion = opinionRepository.findByApiId(opinionId)
             .orElseThrow(() -> new NoEntityException(ErrorCode.ENTITY_NOT_FOUND));
         opinion.delete();
@@ -77,7 +87,8 @@ public class OpinionService {
     }
 
     @Transactional
-    public OpinionOnlyIdResponse completeDeleteOpinion(String opinionId) {
+    public OpinionOnlyIdResponse completeDeleteOpinion(String opinionId, String username) {
+        verifyAuthorOfOpinion(username, opinionId);
         Opinion opinion = opinionRepository.findByApiId(opinionId)
             .orElseThrow(() -> new NoEntityException(ErrorCode.ENTITY_NOT_FOUND));
         opinionRepository.delete(opinion);
@@ -89,5 +100,20 @@ public class OpinionService {
         Opinion opinion = opinionRepository.findByApiId(opinionId)
             .orElseThrow(() -> new NoEntityException(ErrorCode.ENTITY_NOT_FOUND));
         return opinionMapper.entityToOpinionResponse(opinion);
+    }
+
+    private void verifyAuthorOfOpinion(String username, String opinionId) {
+        User user = userRepository.findByUsername(username)
+            .orElseThrow(() -> new NoEntityException(ErrorCode.ENTITY_NOT_FOUND));
+        Opinion opinion = opinionRepository.findByApiId(opinionId)
+            .orElseThrow(() -> new NoEntityException(ErrorCode.ENTITY_NOT_FOUND));
+        if (!user.getUserRoles().stream()
+            .map(UserRole::getRole)
+            .map(Role::getValue)
+            .collect(Collectors.toSet())
+            .contains(RoleEnum.ROLE_ADMIN) &&
+            !opinion.getMemberAuthor().equals(user.getMember())){
+            throw new NotAuthorException(ErrorCode.NOT_OPINION_AUTHOR);
+        }
     }
 }
